+++ app/serializers/activity_comment_serializer.rb
@@ -3,16 +3,12 @@ class ActivityCommentSerializer < ApplicationSerializer
 
   type :comments
 
-  attributes :body, :person_name
+  attributes :body
 
   belongs_to :person
 
   def body
     activity_comment_exhibit = exhibit(object)
     activity_comment_exhibit.body
-  end
-
-  def person_name
-    object.try(:person).try(:to_s)
   end
 end