+++ spec/controllers/api/v2/tasks/comments_controller_spec.rb
@@ -3,9 +3,10 @@ require 'rails_helper'
 RSpec.describe Api::V2::Tasks::CommentsController, type: :controller do
   let(:user) { create(:user_with_account) }
   let(:account_list) { user.account_lists.first }
+  let(:person) { create(:contact_with_person, account_list: account_list).reload.people.first }
   let(:activity) { create(:activity, account_list: account_list) }
-  let!(:resource) { create(:activity_comment, activity: activity, person: create(:person)) }
-  let!(:second_resource) { create(:activity_comment, activity: activity) }
+  let!(:resource) { create(:activity_comment, activity: activity, person: person) }
+  let!(:second_resource) { create(:activity_comment, activity: activity, person: person) }
   let(:id) { resource.uuid }
   let(:parent_param) { { task_id: activity.uuid } }
   let(:parent_association) { :activity }
@@ -14,17 +15,32 @@ RSpec.describe Api::V2::Tasks::CommentsController, type: :controller do
   let(:incorrect_attributes) { nil }
   let!(:not_destroyed_scope) { ActivityComment }
   let(:factory_type) { :activity_comment }
+  let(:correct_relationships) do
+    {
+      person: {
+        data: {
+          type: 'people',
+          id: person.uuid
+        }
+      }
+    }
+  end
+  let(:unpermitted_relationships) do
+    {
+      person: {
+        data: {
+          type: 'people',
+          id: create(:person).uuid
+        }
+      }
+    }
+  end
 
   include_examples 'show_examples'
 
   include_examples 'update_examples'
 
-  context 'create examples' do
-    # The ActivityComment model uses Thread.current[:user] to set it's person_id attribute
-    before { Thread.current[:user] = user }
-    after { Thread.current[:user] = nil }
-    include_examples 'create_examples'
-  end
+  include_examples 'create_examples'
 
   include_examples 'destroy_examples'