+++ spec/serializers/activity_comment_serializer_spec.rb
@@ -8,7 +8,6 @@ describe ActivityCommentSerializer do
   subject { ActivityCommentSerializer.new(activity_comment).as_json }
 
   it { should include :body }
-  it { should include :person_name }
   it { should include :created_at }
   it { should include :updated_at }
   it { should include :updated_in_db_at }
@@ -19,16 +18,9 @@ describe ActivityCommentSerializer do
     end
 
     it { should include :body }
-    it { should include :person_name }
     it { should include :created_at }
     it { should include :updated_at }
     it { should include :updated_in_db_at }
-  end
-
-  describe '#person_name' do
-    it "returns the person's name" do
-      expect(subject[:person_name]).to eq activity_comment.person.to_s
-    end
   end
 
   describe '#body' do